requirements.txt
# Core dependencies
streamlit==1.31.0
numpy==1.24.3
requests==2.31.0

# Audio processing
sounddevice==0.4.6
soundfile==0.12.1
ffmpeg-python==0.2.0

# Speech recognition and synthesis
openai-whisper==20231117
azure-cognitiveservices-speech==1.31.0

# Deep learning
torch==2.1.0
torchaudio==2.1.0

# Optional but recommended for better performance
python-dotenv==1.0.0
